Edit/Edit/requirements.txt
Django==1.11.11
pandas==0.22.0
djangorestframework==3.7.7